// ts/packages/trellis/sdk/_generated/health/schemas.ts
export const NotFoundErrorDataSchema = {
  "properties": {
    "context": { "additionalProperties": true, "type": "object" },
    "id": { "minLength": 1, "type": "string" },
    "message": { "type": "string" },
    "resource": { "minLength": 1, "type": "string" },
    "traceId": { "type": "string" },
    "type": { "const": "NotFoundError", "type": "string" },
  },
  "required": ["type", "resource", "id", "message"],
  "type": "object",
} as const;